# docs/swagger.yaml
definitions:
...
  models.Loai:
    properties:
      hinh:
        type: string
      ma_loai:
        type: integer
      mo_ta:
        type: string
      ten_loai:
        type: string
      ten_loai_alias:
        type: string
    type: object
  models.NhaCungCap:
    properties:
      dia_chi:
        type: string
      dien_thoai:
        type: string
      email:
        type: string
      logo:
        type: string
      ma_ncc:
        type: integer
      mo_ta:
        type: string
      nguoi_lien_lac:
        type: string
      ten_cong_ty:
        type: string
    type: object
...
  models.TrangThai:
    properties:
      ma_trang_thai:
        type: integer
      mo_ta:
        type: string
      ten_trang_thai:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /loai/:
    get:
      consumes:
      - application/json
      description: Hiển thị tất cả các loại hàng hóa
      produces:
      - application/json
      responses: {}
      summary: Lấy ra tất cả các loại hàng hóa
      tags:
      - Loại
    post:
      consumes:
      - application/json
      description: Thêm loại hàng hóa
      parameters:
      - description: Loai data
        in: body
        name: loai
        required: true
        schema:
          $ref: '#/definitions/models.Loai'
      produces:
      - application/json
      responses: {}
      summary: Thêm Loại
      tags:
      - Loại
  /loai/{ma_loai}:
    delete:
      consumes:
      - application/json
      description: Xóa loại hàng hóa theo mã hàng hóa
      parameters:
      - description: Maloai
        in: path
        name: ma_loai
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Xóa loại hàng hóa
      tags:
      - Loại
    get:
      consumes:
      - application/json
      description: Hiển thị loại hàng hóa theo mã hàng hóa
      parameters:
      - description: MaLoai
        in: path
        name: ma_loai
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Lấy loại hàng hóa theo mã hàng hóa
      tags:
      - Loại
    put:
      consumes:
      - application/json
      description: Chỉnh sửa loại hàng hóa theo mã hàng hóa
      parameters:
      - description: MaLoai
        in: path
        name: ma_loai
        required: true
        type: integer
      - description: Loai data
        in: body
        name: loai
        required: true
        schema:
          $ref: '#/definitions/models.Loai'
      produces:
      - application/json
      responses: {}
      summary: Sửa loại hàng hóa theo mã hàng hóa
      tags:
      - Loại
  /nhacungcap/:
    get:
      consumes:
      - application/json
      description: Hiển thị tất nhà cung cấp cho hàng hóa
      produces:
      - application/json
      responses: {}
      summary: Hiển thị nhà cung cấp
      tags:
      - Nhà Cung Cấp
    post:
      consumes:
      - application/json
      description: Thêm nhà cung cấp cho hàng hóa
      parameters:
      - description: NhaCungCap data
        in: body
        name: nhacungcap
        required: true
        schema:
          $ref: '#/definitions/models.NhaCungCap'
      produces:
      - application/json
      responses: {}
      summary: Tạo nhà cung cấp
      tags:
      - Nhà Cung Cấp
  /nhacungcap/{ma_ncc}:
    delete:
      consumes:
      - application/json
      description: Xóa nhà cung cấp theo mã nhà cung cấp
      parameters:
      - description: MaNCC
        in: path
        name: nhacungcap
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Xóa nhà cung cấp theo mã nhà cung cấp
      tags:
      - Nhà Cung Cấp
    get:
      consumes:
      - application/json
      description: Hiển thị nhà cung cấp theo mã nhà cung cấp
      parameters:
      - description: MaNCC
        in: path
        name: nhacungcap
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Hiển thị nhà cung cấp theo mã nhà cung cấp
      tags:
      - Nhà Cung Cấp
    put:
      consumes:
      - application/json
      description: Chỉnh sửa nhà cung cấp theo mã nhà cung cấp
      parameters:
      - description: MaNCC
        in: path
        name: nhacungcap
        required: true
        type: integer
      - description: NhaCungCap data
        in: body
        name: nhacungcap
        required: true
        schema:
          $ref: '#/definitions/models.NhaCungCap'
      produces:
      - application/json
      responses: {}
      summary: Chỉnh sửa nhà cung cấp theo mã nhà cung cấp
      tags:
      - Nhà Cung Cấp
...
  /trangthai/:
    get:
      consumes:
      - application/json
      description: Hiển thị trạng thái
      produces:
      - application/json
      responses: {}
      summary: Hiển thị trạng thái
      tags:
      - Trạng thái
    post:
      consumes:
      - application/json
      description: Thêm trạng thái
      parameters:
      - description: TrangThai data
        in: body
        name: trangthai
        required: true
        schema:
          $ref: '#/definitions/models.TrangThai'
      produces:
      - application/json
      responses: {}
      summary: Thêm trạng thái
      tags:
      - Trạng thái
  /trangthai/{ma_trang_thai}:
    delete:
      consumes:
      - application/json
      description: Xóa trạng thái theo mã trạng thái
      parameters:
      - description: MaTrangThai
        in: path
        name: ma_trang_thai
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Xóa trạng thái theo mã trạng thái
      tags:
      - Trạng thái
    get:
      consumes:
      - application/json
      description: Hiển thị trạng thái theo mã trạng thái
      parameters:
      - description: MaTrangThai
        in: path
        name: ma_trang_thai
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Hiển thị trạng thái theo mã trạng thái
      tags:
      - Trạng thái
    put:
      consumes:
      - application/json
      description: Chỉnh sửa trạng thái theo mã trạng thái
      parameters:
      - description: ma_trang_thai
        in: path
        name: ma_trang_thai
        required: true
        type: integer
      - description: TrangThai data
        in: body
        name: trangthai
        required: true
        schema:
          $ref: '#/definitions/models.TrangThai'
      produces:
      - application/json
      responses: {}
      summary: Chỉnh sửa trạng thái theo mã trạng thái
      tags:
      - Trạng thái
swagger: "2.0"